requirements.txt
pillow==12.2.0
openai==2.31.0
python-dotenv==1.2.2
requests==2.33.1
beautifulsoup4==4.14.3
google-genai==1.73.1
